Split the interpreter operations into multiple files
This allows the individual operations to have smaller source code files, which
should be easier to understand.
