--- mu/Interpreter.hpp
@@ -1,4 +1,0 @@
-#pragma once
-
-void Add();
-void Subtract();
