+++ mu/Interpreter/Subtract.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+void Subtract();
